+++ src/Websites/Websites/Cmdlets/Certificates/ImportAzWebAppKeyVaultCertificate.cs
@@ -66,7 +66,16 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
                 {
                     kvid = KeyVaultName;
                 }
-           
+                string keyvaultperm;
+                keyvaultperm = CmdletHelpers.CheckServicePrincipalPermissions(this.ResourcesClient, this.KeyvaultClient, this.ActiveDirectoryClient, kvresourcegrpname, KeyVaultName);
+                var lnk = "https://azure.github.io/AppService/2016/05/24/Deploying-Azure-Web-App-Certificate-through-Key-Vault.html";
+                if ((keyvaultperm != "Get") & (keyvaultperm != "get"))
+                {
+                    WriteWarning("Unable to verify Key Vault permissions.");
+                    WriteWarning("You may need to grant Microsoft.Azure.WebSites service principal the Secret:Get permission");
+                    WriteWarning(string.Format("Find more details here: '{0}'", lnk));
+                }
+
                 Certificate kvc = null;
                 var certificate = new Certificate(
                     location: location,